fix(android): 编译错误修复——首次成功构建出 APK（第十一轮）
经 assembleDebug 编译暴露并修复 5 类错误（此前仅靠静态检查，无法发现）：
- ChatActivity: 轮询调 refreshMessages() 但方法签名已改 boolean → refreshMessages(false)
- EditPostActivity: loadCategories 用了未定义的 gson/log → 补字段与 log 方法
- PostDetailActivity: postMultipart 需 Map<String,Object>（传的是 String map）→
  multipart 分支改用独立 Map<String,Object>
- PostsFragment: setPadding 需 int（applyDimension 返回 float）→ 显式 (int)
- CommentsAdapter: Reply 头像 onUserClick 需经 OnReplyActionListener 转发，
  NestedRepliesAdapter 接口补 onUserClick + 两处匿名实现补齐

构建结果：app-debug.apk（7.8MB）构建成功
（SDK F:/AndriodSDK + JDK 21 + Gradle 9.3.1，详见记忆）

+++ app/src/main/java/com/schoolforum/app/ui/post/PostsFragment.java
@@ -303,10 +303,10 @@ public class PostsFragment extends Fragment implements PostsAdapter.OnPostClickL
         chip.setTextSize(13);
         chip.setMinWidth(0);
         chip.setMinHeight(0);
-        chip.setPadding(android.util.TypedValue.applyDimension(android.util.TypedValue.COMPLEX_UNIT_DIP, 12, getResources().getDisplayMetrics()),
-                android.util.TypedValue.applyDimension(android.util.TypedValue.COMPLEX_UNIT_DIP, 4, getResources().getDisplayMetrics()),
-                android.util.TypedValue.applyDimension(android.util.TypedValue.COMPLEX_UNIT_DIP, 12, getResources().getDisplayMetrics()),
-                android.util.TypedValue.applyDimension(android.util.TypedValue.COMPLEX_UNIT_DIP, 4, getResources().getDisplayMetrics()));
+        chip.setPadding((int) android.util.TypedValue.applyDimension(android.util.TypedValue.COMPLEX_UNIT_DIP, 12, getResources().getDisplayMetrics()),
+                (int) android.util.TypedValue.applyDimension(android.util.TypedValue.COMPLEX_UNIT_DIP, 4, getResources().getDisplayMetrics()),
+                (int) android.util.TypedValue.applyDimension(android.util.TypedValue.COMPLEX_UNIT_DIP, 12, getResources().getDisplayMetrics()),
+                (int) android.util.TypedValue.applyDimension(android.util.TypedValue.COMPLEX_UNIT_DIP, 4, getResources().getDisplayMetrics()));
         android.widget.LinearLayout.LayoutParams lp = new android.widget.LinearLayout.LayoutParams(
                 android.widget.LinearLayout.LayoutParams.WRAP_CONTENT,
                 android.widget.LinearLayout.LayoutParams.WRAP_CONTENT);
